fix: Limit departures widget backend to send up to 15 departures to the client
We ran into a rare case on certain bus e-ink screens at very busy stops,
where a large number of upcoming departures (60+) were sent to the client.

This caused the client's row-fitting logic to make too many react state updates and crash
the page, since it removes one departure from the list at a time until they fit.

(Until we are able to pre-calculate the total rendered height of a given list of
departures, this is the only way to fit rows in the available space.)