chore: fix NPM broken peer dependencies
The NPM that came with Node 14 would sometimes allow deps to resolve
even when packages' peer dependencies had incompatible versions
(either among each other, or with the dependencies as specified in
`package.json`). This was the case for a couple of our deps, so once we
upgrade to the new lockfile format, `npm install` realizes the issue
and won't proceed unless we provide a "use the old broken behavior"
flag. Since the required changes appear minor, we'll just fix them.

* `@types/jest` was at some point left behind on version 26 despite
  `jest` itself being upgraded to 27, so we bring it up to date.

* `ts-loader` was at some point bumped from 8 to 9 by Dependabot,
  despite this version technically requiring Webpack 5 while we are
  still on 4. Downgrading it doesn't appear to cause any issues.