chore: update to latest config version
The new version of the departures widget config moves route-direction
filtering inside a more general "filters" field, to allow adding other
filters ("max minutes" for busway screens). The existing `filter` field
is unused and set to `null` on all current screens, so we don't have to
worry about migrating, but we did already have some code to apply them
if any had been specified, so that needs to be updated.