fix: refresh-rate-offset should not be required
This was causing a crash / blank page on the "all screens" views, which
do not provide a refresh rate offset in the top-level HTML element.
From the immediate context it seemed like this value must have been
required, since it was passed into `parseInt` which would return a
`NaN` if undefined — but it appears later code was tolerant of this.

This changes the value to specifically default to `0` at the time it is
loaded from the HTML.