refactor: dep injection in departures generator
The main function of the Departures candidate generator,
`departures_instances`, accepted a dependency injection for a function
that was part of its own implementation, which made testing the logic
within that function difficult.

Here we smooth out the "injection surface" a bit so the function being
injected is consistently the lower-level `fetch` function that performs
the API calls, and refactor the tests accordingly. Some factory
functions are introduced to simplify the implementation of multiple
similar tests.

In addition:

* Some typespecs for relevant functions are added.
* A missing test is added for the candidate generator's "no service"
  behavior for bus e-ink screens.