dev: migrate to ESLint and resolve lint errors
TSLint was deprecated in favor of ESLint several years ago:
https://github.com/palantir/tslint/issues/4534

Additionally, running the linter has not been part of CI, so (much like
type-checking and formatting) we've accumulated many lint issues in the
code over time.

Rules where we have a huge number of outstanding issues (primarily
`no-explicit-any` and `prop-types`) are set to `warn` for the time
being, so they are still visible but won't fail CI. Remaining issues
have been cleaned up.