refactor: "route abbreviations" for route pills
"Route abbreviations" (formally named in this changeset) are short text
codes that stand in for the names of routes which would otherwise be
represented by a "mode icon" in route pills. Currently this only applies
to Commuter Rail, though could conceivably apply to ferry in the future.

Previously, whether a "route abbreviation" was used was controlled at
the time the route pill was serialized — the result from the client's
perspective would either be an "icon pill" with the mode icon or a "text
pill" with the abbreviation. However, the "Later Departures" feature
requires that the client know both of these pieces of information at
once, since which one is used depends on context that is not known until
the widget is rendered.

Accordingly, this adds a field `route_abbrev` to the serialized form of
icon and text pills (needed for text pills as well because Commuter Rail
departures where the train has a track assignment are represented with
a "TRx" text pill). The client's `RoutePill` component gains a prop that
determines whether to use the abbreviation in the pill data. If true, it
renders a text pill with the abbreviation as the content.