fix: Later Departures header pills overflowing
When the original approach for Later Departures was changed to use a
carousel that renders N copies of the component, we missed that this
broke the auto-trimming feature. An easy way to fix this is to have the
carousel only consist of the departure rows themselves, with the header
pills being fixed.

This includes a style update to ensure we only display the "dividers"
between departure rows when those rows are direct children of a section,
rather than inside a Later Departures component.