dev: fix Phoenix assets watcher not exiting
Phoenix expects watcher processes to exit when it sends an EOF to their
stdin, but tools like Webpack and `tsc` don't have this behavior by
default and will keep running in the background, piling up over time.
`exit_on_eof` is a package developed specifically to bridge this gap.

(Webpack actually has a `--watch-options-stdin` flag for this, but since
`tsc` lacks such an option, we would need to use the package anyway.)