fix: inconsistent departure sorting
When displaying departures, we show the arrival time if there is one,
and the departure time if not. When fetching departures, however, we
were sorting by departure time if there was one, and arrival time if
not (unless also fetching scheduled stop times, in which case sorting
used the same logic as displaying!). This created some situations where
departures would appear "out of order".