fix: sections not always respecting layout `max`
The `max` field is intended as a hard limit: a higher number of rows
should never be displayed (aside from Later Departures). However, if
the layout already fit on the screen without overflowing on the initial
render, we would never trim sections at all, including to their max.

This makes the initial conversion of a section to a "folded" section
also perform the trim-to-max.