fix: prevent crash on empty list
Refactors `select_direction_and_route/1` to use pattern mattern more
heavily and handle cases when the passed argument is an empty list.

This handles the case when the direction and route for the informed
entities of an alert are empty or otherwise filtered down to an empty
list (e.g. only contain JFK). Additionally this is perhaps slightly more
idiomatic Elixir code.