fix: only set cached=true when a 304 is received
Previously our telemetry logging would show `cached=true` whenever we
had a possible value for a request in the cache **not** when we actually
used the cache. This was misleading and is corrected in this change.

Being able to return a meta value in a `Screens.Telemetry` span
required a new function that expects the same return value as
`:telemetry.span/3` which is a tuple where the second element is a map
of metadata to include in the `*.stop` event.