feat: Filtering for in-memory Alerts cache
Supports filtering the Alerts in the in-memory cache on route,
route_type, direction_id, stop, and activity.

Route, stop, route type, and activity filters are lists of values where
_any_ of these values need to be present in at least one of the informed
entities of an alert to match.

If no activity filters are provided the default activity filter of
BOARD, EXIT, or RIDE is applied.

Direction ID can only be filtered on one value at a time because filtering
on multiple values is the same as removing the filter altogether.

If no filters are passed no filtering is done including no filtering on
the default activity filter.