fix: apply filters to alerts cache correctly
The alerts cache was missing a step that's present in the V3 API which
meant that filters were not being applied to the list of all alerts
correctly producing the wrong results.

This change adds a test suite to verify that filters work as expected
and adds the additional step of expanding built matchers so partial
matches are considered a match.