fix: guard against inspector code crashing screens
On real Bus E-ink hardware we saw a mysterious "DOM exception 12" crash
that seemed to be caused by the inspector changes. Though the message
sending and listening should have already been a no-op when the screen
is not "framed", this adds a check for whether the screen is "framed"
by the inspector specifically, and otherwise doesn't even attempt to
send/listen. This should sidestep any issues with e.g. a screen vendor
rendering our pages inside a frame.