fix: handle unscheduled commuter rail departures
The commuter rail departures widget made the assumption that there would
never be a departure that had a prediction but no matching schedule. We
have observed that this isn't always the case.

Adds handling of a `nil` schedule to the commuter rail departures
widget. The only thing relying on a present schedule in
`serialize_departure/5` was calculating if a departure was delayed. If
there is no schedule we do not flag the departure as delayed; otherwise,
the delayed calculation remains unchanged.