fix: always use API_V3_URL environment variable
Our ECS tasks all set a `API_V3_URL` pointing to the desired V3 API
environment but we were ignoring that value and instead deriving the V3
API URL based on `ENVIRONMENT_NAME`. This is confusing.

With this change the URL used for the V3 API is always fetched from the
environment as `API_V3_URL` (unless `MIX_ENV=test`).

This has the added benefit of picking up the changes in https://github.com/mbta/devops/pull/2122