chore: remove bespoke routes-at-stop retry logic
Some time ago (36f9f83) we changed this specific data fetching function
so it "doesn't trust" the V3 API if it says there are no routes serving
a stop, and tries again up to 3 times.

Some more-recent time ago (297c302), as part of a refactor of this
function, we added a warning when a retry occurs, so we'd get some idea
whether this logic was still useful.

One month later, we have not logged this warning in any deployed
environment, so it seems safe to remove.