feat: switch and parallel running for DUP logic
Introduces the concept of a "variant", which is an alternate candidate
generator for an app ID. Also introduces a variant for DUPs which will
be used to build and test the new departures logic. For now, this only
generates a set of placeholder elements.

Variants can be enabled via:

* Adding a `variant=NAME` query param to a screen URL. The resulting
  screen page will fetch and display data for only that variant.

* Enabling the "variant switcher" in the admin Inspector. This allows
  toggling between variants instantly without refreshing data; the data
  hook uses a special `all` value to request and hold onto the data for
  all variants at once.

Additionally, `ScreenData` functions have an option to run and serialize
all variants that exist for the app ID in parallel, in the background,
without waiting for or doing anything with the results. This means we
can have all existing screens constantly running the new logic at every
step of the implementation, giving us more confidence that the final
rollout would go smoothly. For now, we only use this option on "normal"
screen requests (not pending, not simulation).