feat: suppress periodic audio in inspector
Avoids being jumpscared by a periodic audio readout that also can't be
silenced, short of closing the tab or navigating away. To replace the
missing feedback on whether periodic audio is configured, also adds a
line to the inspector audio controls to report the readout parameters
that would have been used.