refactor: fetch parents/children with stops
Some stop functions were already using this relationship by effectively
hand-rolling the relevant bits of the parsing. We can simplify by fully
parsing the parent/child relationships in the main stop `Parser`, and
having the special-purpose `Stop` functions delegate to a main `fetch`
function as we do with other resources.

Since we may be "starting from" either a parent or a child, we need a
way to encode the complete data without recursing infinitely through
this bidirectional link. For this purpose we introduce `:unloaded` as a
special value indicating a relationship was not populated, similar to
`Ecto.Association.NotLoaded`. This allows ending the recursion once we
know we have all the data. It also allows the relationships to not be
fetched in the first place, for functions that don't need this data or
in contexts where stops are being parsed as a relationship on another
resource.