chore: look up headway values by stop
Instead of manually configuring a "headway key" for every departures
section, we'd like to be able to translate stop IDs (which departures
already must be configured with) into headway keys.

Currently the mapping of "headway key" to stop IDs only exists
implicitly in the PA/ESS sign configuration in `realtime_signs`, and
then only for stations that have PA/ESS signs. So for now, we need to
encode this information ourselves within Screens.