fix: elevator closures sorting instability
The previous code mutated the `stations` prop when sorting, and also did
not implement a stable sort function[^1]. As a result, the closure list
would be reversed each time a data request occurred. In the common case
where the widget has two pages, this resulted in certain closures never
being shown, since they would always be on the "other" page.

[^1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/sort#description