dev: fix module resolution issue with `util`
We have a module `util` in a directory also named `util`. This happens
to work fine for all current packages, but [Knip](https://knip.dev/) is
unable to resolve this specific module, complaining both that `util` is
"unused" and that all references to `util` are "unlisted dependencies".
This does not appear to be a bug in the library; it seems more likely
our naming here is going against some convention of JS module naming and
doing something that isn't universally supported. In any case, it's easy
to change.