refactor: introduce structs for departure sections
Sections passed to the `Departures` widget for serialization were
previously plain maps with a `type` field. In addition to lacking any
compile-time verification of fields or other struct conveniences, this
had some confusing overlap with the serialized form of sections, which
are maps with a `type` field (of the same value, but with the rest of
the fields being different!).

* Introduce `Departures.XSection` structs for each type of section the
  widget supports.

* Remove the plain-map indirection of "notice rows". Instead of
  `%{text: FreeTextLine.t()}` they are just `FreeTextLine.t()`.

* Does not remove some existing instances of `%{type: :x_section}` in
  the DUP departures candidate generator; these are part of an "internal
  API" of the module and are neither real Departures widget sections nor
  the serialized form.