chore: remove unused/broken BlueBikes widget
This feature has been unused since the original Orange Line "surge" it
was developed for in mid-2022, and non-working since July 2023, when the
data fetching was commented out due to persistent crashes (see 9211202).

We should avoid leaving code committed that will never be executed
and/or features available in the admin UI that don't work. Should we
choose to prioritize this feature again, the old code lives on in the
history of the repo for reference.