refactor: remove separate Hackney pool for ExAws
Though this is unlikely to be causing any issues, it also seems very
unnecessary. Over the past 30 days we've seen at most 2 concurrent
connections out of a possible 50 in this pool (and 1 out of 50 for the
default pool, which ExAws will use after this change). This aligns with
how often we'd expect to be calling AWS APIs during normal operations.