fix: incorrect boldness in elevator outages
For some elevators with in-station redundancy, the text "Accessible
route available" would incorrectly be shown in the bold font reserved
for more noteworthy redundancy categories.

The reason is that we were expecting the redundancy category to be
exactly `2`, but some elevators had categories like `2A` or `2B`. Since
the exiting summary for all category-2 elevators is "Accessible route
available", this resulted in screens still showing the correct phrasing,
but more boldly than intended.