feat: initial support for "solo" Pre-Fares
A Pre-Fare solo is, for most practical purposes, just the right screen
of a Pre-Fare duo. This adds a configuration option to formalize this,
plus a tweak to the header so the stop name and time are both displayed
on the right (only) side.

Widgets otherwise behave identically, except that alerts which would
take over both screens of a duo will not appear at all (not egregious
compared to the current behavior, where most of the useful information
is displayed on the nonexistent left screen, leaving only a vague "Seek
alternate route" message on the visible one). Full alert support for
solo screens will be implemented in an upcoming task.

Key technical changes:

* `CandidateGenerator.screen_template` now accepts an argument which is
  the screen configuration, allowing the generated template to vary per
  individual screen.

* Simplify the `Viewport` component out of existence.