feat: alert widget support for solo Pre-Fares
The existing alert widget "hierarchy" of dual screen > single screen >
flex zone does not make sense on Pre-Fares that only have one screen to
work with. We "adjust" placement for solo screens by moving all alerts
down one level in this hierarchy (with some exceptions):

* Alerts that would have been dual-screen are single-screen.
* Alerts that would have been single-screen are (mostly) flex-zone.
* Alerts that would have been in the flex zone stay there.

This reorganizes some of the tests for this widget and adds assertions
for solo screens to existing test cases. Initially most of these are
placeholders whose main purpose is to ensure we don't crash in any
known scenarios.