test: fix nonexistent stop for ReconstructedAlert
This test case used `place-swnxg` as the informed stop of an alert
(possibly a typo of `place-dwnxg`, Downtown Crossing), which was not
included in the tagged stop sequences of the test widget, despite the
intent of this case being to test a "downstream" delay alert.

Fixing this requires a change to the assertion, since a more-specific
map of route data is generated (the full-screen version of the widget
would say "Attention, riders to Forest Hills" rather than "Attention OL
riders").