dev: remove `babel-loader` from Dependabot group
This package is a Webpack plugin and does not align with Babel's major
releases, resulting in some odd PRs like #2487 where Dependabot tries
to update to a new patch version of `@babel/core` and a new _major_
version of `babel-loader`. With this line removed, the latter is sorted
into the `webpack` group, due to the existing `*-loader` pattern.