feat: "here" station closures for Pre-Fare solo
The code previously assumed a station closure alert "here" (at the
screen's station) would only be presented as a single-screen takeover
if it was a closure of a single route at a transfer station; if an
alert eliminated all service at the station, it would be a dual-screen
takeover. This assumption is no longer valid with the addition of
the "solo" template, which uses a single-screen takeover for anything
that would have been a dual-screen takeover on a "duo".

This change enables such "full" station closures to be displayed as a
single-screen takeover, with a disruption diagram and text matching the
dual-screen version.