fix: don't try to compile Credo check in prod
cf0a573 made Credo a dev/test-only dependency, but we have a custom
Credo check in the `lib` directory which Elixir will therefore try to
compile even in the `prod` environment, failing because Credo is not
loaded there.

We could do something with `elixirc_paths` here (e.g. moving the module
to a top-level `lib_dev` directory and only compiling that directory in
dev/test), but this change goes the route of simply wrapping the module
definition in a `Code.ensure_loaded?` conditional.