refactor: fetch and parse Facility resources
* Introduce a `Facility` struct and generic fetch functions for it.
  Allows replacing the narrower `fetch_stop_for_facility` function with
  fetching the facility and reading its `stop` relationship.

* Parse facilities related to alerts as the complete `Facility` instead
  of a map of ID and short name, potentially including relationships of
  the facilities as well. Allows replacing the special-purpose
  `fetch_elevator_alerts_with_facilities` with the normal `fetch`,
  since related facilities are now always included.

* Fix a minor bug in Stop parsing where `connecting_stops` being
  unloaded was represented as `nil`, rather than `:unloaded` as the
  typespec suggested.

For now only the `excludes-stop` facility property is parsed. This plus
the `stop` relationship will ultimately let us determine the child stop
IDs affected by a facility alert (e.g. the specific platforms within a
station that may be inaccessible if a given elevator is offline), which
is a necessary component of determining whether a given elevator outage
is "downstream" of a screen location.