refactor: DRY up V3 API response parsing
Some patterns were repeated across many of our resource parsing modules,
including an implicit "API" for passing around included resources. This
centralizes the shared logic in a `V3Api.Parser` module.

Incidentally fixes a crash that occurred when trying to resolve a
relationship that referred to a resource in the response's `data` (not
`included`). This could occur e.g. when a response had a `stop` with a
parent station and `parent_station.child_stops` was included (since this
relationship would refer back to the stop itself).