refactor: use facility stop in elevator closures
Generating elevator screen closures requires, among other things, the
ID and name of the elevator's station.

Previously the ID was obtained from the `stop` field of the same
`informed_entity` that referenced the elevator (assuming the alerts
feed would have to populate this, given "facilities" are a non-standard
extension), and the name from a separate API fetch that collected the
names of all parent stations in the system.

We can now populate the facilities associated with an alert and the
stops associated with those facilities as part of the alert fetch, so
this information can be gotten directly from the Stop resources. In
particular this removes the need for the separate "station name map" to
be fetched, passed around, and stubbed in tests.

Fetching alerts with the full Facility relationship graph populated
(`include_all?: true`) will also enable more easily determining which
child stop IDs within a station are affected by an elevator closure,
which we need to enable "downstream" closure summaries.