ci: fix caches never being populated on main
Because we only had CI running on `pull_request`, it was never actually
running on the main branch, which is the only branch whose caches other
branches are allowed to use. This resulted in builds on feature branches
slowing down as the main cache became progressively less usable due to
dependency changes.