fix: crash in elevator screen logic
Due to a quirk of our data, shuttle bus stops that are considered child
stops of the parent stations they "replace" are incorrectly implied to
be served by elevators within the parent station. As a result, shuttle
bus route patterns (which have some stops with no parent station) ended
up in our "downstream" elevator closures logic (which assumed all stops
would have a parent station).

We did not need to match on the full structure of the stop in this case,
so it can be simplified.