feat: DUP secondary departures no-data fallback
When no sections were configured for DUP secondary departures, the
secondary rotation would be a copy of the primary departures. This
extends that behavior to also occur when sections *are* configured but
there are currently no departures to display. The primary use case is
DUPs that display subway (primary) and bus (secondary) departures, where
the relevant bus stops have no service on weekends.

Also restructure logic to reduce redundant data fetching, and add some
typespecs.