feat: support footers with no stop ID
All screen types that support `Footer` configuration now support having
no stop ID specified, for screens that are not associated with a single
stop. The "more info" URL displayed in this case is simply `mbta.com`.

Includes minor refactors to remove unneeded "variability" of values that
were always the same.