refactor: move name fetch out of LocationContext
The `home_stop_name` field of `LocationContext` was fetched on all
screens that use this module, but only used in one specific case when
building ReconstructedAlerts for Pre-Fares. Move this field out of
LocationContext and onto ReconstructedAlert itself, alongside the
`informed_station_names` field it is paired with in the widget logic.

Incidentally fixes a bug where audio equivalence would be slightly wrong
when the home station was Massachusetts Ave, as this was translated into
"Mass Ave" in one of these fields but not in the other.