feat: support multi-stop alerts configuration
These are the minimal changes required to support Bus E-ink screens
(only) being configured to display alerts for multiple stops.

Presentation may not be ideal due to conditions involving the screen's
"home stop" always being considered false (as there is no single "home
stop"), but since we have exactly one use case for this feature and
don't expect that to change soon, the amount of effort applied here is
appropriate to the impact.