fix(admin): carousel issue when refresh disabled
When "Enable refresh interval" was unchecked in the Inspector, Later
Departures carousels on Sectionals would rotate extremely fast, due to
the `null` refresh rate being plugged into a calculation that resulted
in the rotation delay being 0.

This setting was only ever intended to affect whether screen data
automatically refreshes, not calculations that depend on the refresh
rate, so this scopes it down to just the API data fetching hook.