fix: Pre-Fare flex zone crash with departures
At transfer stations, having departures enabled could force an "inside"
shuttle or suspension alert into the flex zone, which we did not have a
relevant function clause for. To resolve this crash in the short term,
reuse the presentation of "boundary" shuttle/suspension alerts, which
does not actually have any elements specific to being on the boundary.