fix: incorrect pattern match in device monitor
This pattern was from an earlier version of the code that used the
`cas: true` option for `Memcache.add`, which changes the shape of the
`:ok` return value. Fortunately this didn't cause any problems since
it caused the DeviceMonitor to crash and restart one time only (when
the key was initially stored) in a situation where it would have done
nothing anyway.