fix: device monitor logging interval inconsistency
Multiple instances of the app with staggered startup times could result
in running logging with an interval shorter than the intended 5 minutes
(for example we could get into a cycle of logging a 4-minute span, then
the following 1 minute, then the following 4 minutes, etc.)

To make this more consistent and less fragile, ideally the monitor
should be able to check the stored "last run" timestamp independently
of deciding whether to update it (and begin a new run). Therefore this
splits the store's `get_and_update/2` into separate `get/1` and
`set/3`, exposing the version tag used to check for conflicts.

With this change, we should log exactly once per clock-aligned 5-minute
interval (at minute 0, 5, 10, etc.)