refactor: DupAlert layout selection logic
The layout of a DUP alert is determined by certain rules. These were
implemented by essentially pre-applying them to every combination of
inputs to create a TSV-like "lookup table", which was then parsed into
a more usable form. This was over-complicated given how straightforward
the rules themselves are (and how irrelevant much of the data in the
table was as a result). Expressing them directly as code results in a
more concise and legible implementation, and makes it easier to change
the rules in the future.