feat: single-tracking messaging for DUP alerts
Most widgets technically already support single-tracking alerts since
these have an effect of `DELAY`. However, when the cause of a delay is
single-tracking specifically, we want to make this info more prominent,
including showing the alert at severity levels we otherwise wouldn't.
This updates the handling of DUP alerts accordingly.