fix: DUP subway headways appearing in bus sections
When a DUP was configured with secondary bus departures, and on
the "boundary" of an active alert for the subway route in primary
departures, the secondary sections would incorrectly display headway
messages for the subway route instead of the (still present) bus
departures.

This issue was introduced in 6986ffb and caused by an incorrect
implementation of headway key lookups, where if a station didn't have
any transfers to other subway routes, it was assumed that the station
served no other routes at all (including e.g. bus routes).

This updates the behavior of headway lookups to be more conservative,
fixing the issue but with the side effect that sections not configured
with an explicit route ID filter will never show headways.