fix: incorrect location context for DUP alerts
To determine whether a DUP was inside/outside/on the boundary of a
subway disruption, we were incorrectly considering the stop sequences
of *all* route patterns, including atypical ones used for disruptions.
This could result in alerts thinking a screen was "inside" a disruption
it was actually on the boundary of, and displaying e.g. "No BL trains"
instead of "No BL trains to Wonderland".

This fixes location context fetching to only consider canonical stop
sequences on DUPs, when possible. When not possible (relevant mostly to
the Silver Line DUP at WTC), we use the most typical sequences.