fix: tolerate invalid config in no-data message
The logic for directional no-data messages added in b2b1ca40 encountered
a crash when screens were mistakenly configured with a `direction_id`
of `null`, when the correct value to use for no direction filtering is
`"both"`. This restructures the logic to check specifically for a valid
direction ID, and otherwise show the generic message.